Rename lp_diag.lpdiag → lp_diag.cli
