+++ message_ix/tests/tools/test_lpdiag.py
@@ -1,7 +1,7 @@
 import pytest
 
+from message_ix.tools.lp_diag.cli import main
 from message_ix.tools.lp_diag.lp_diag import LPdiag
-from message_ix.tools.lp_diag.lpdiag import main
 
 
 def test_lpdiag(test_data_path):
